% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDatasource.R
\name{addDatasource}
\alias{addDatasource}
\title{addDatasource}
\usage{
addDatasource(jsonOutputPath, dataName = NULL, datasourcePath = NULL,
  datasourcesFilename = "datasources.json",
  additionalParam = list(cdmVersion = 5))
}
\arguments{
\item{jsonOutputPath}{Path of the Json files generated by \code{exportToJson}.}

\item{dataName}{Name of Achilles report. Default is the base folder of \code{jsonOutputPath}.}

\item{datasourcePath}{Path where datasource file will be saved. Default is one folder above the \code{jsonOutputPath}}

\item{datasourcesFilename}{Name of the file where the datasource is located or stored. Default is "datasources.json".}

\item{additionalParam}{A R list specifying which additinal parameters to write to the datasource object. Default is \code{list(cdmVersion=5)}.}
}
\value{
none
}
\description{
\code{addDatasource} adds a data source to the datasource.json file.
}
\details{
Used to update the datasources file with the reference to a specified datasource.
This makes the new datasource findable for OHDSI tools.
If the datasources file exists, the data source will be added to the file.
If the datasources file does not exist, a new file wil be initialized with the specified data source.
}
\examples{
jsonOutputPath <- "your/output/path"
connectionDetails <- createConnectionDetails(dbms="sql server", server="yourserver")
exportToJson(connectionDetails, cdmDatabaseSchema="cdm5", outputPath=jsonOutputPath)
addDatasource(jsonOutputPath, "your_data_name")

}

