% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{conceptHierarchy}
\alias{conceptHierarchy}
\title{execution of concept hierarchy creation}
\usage{
conceptHierarchy(connectionDetails, vocabDatabaseSchema,
  oracleTempSchema = vocabDatabaseSchema,
  resultsDatabaseSchema = vocabDatabaseSchema, cdmVersion = "5")
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is vocabDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all temporary tables to be managed. Requires create/insert permissions to this database.}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is vocabDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{cdmVersion}{Define the OMOP CDM version used:  currently support only "5".  Default = "5"}
}
\value{
nothing is returned
}
\description{
\code{conceptHierarchy} executes script to create the concept_hierarchy table.
}
\details{
\code{conceptHierarchy} executes script to create the concept_hierarchy table in the results schema, to be used by Atlas for treemap displays.
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="myserver")
  conceptHierarchy <- conceptHierarchy(connectionDetails, resultsDatabaseSchema="scratch", vocabDatabaseSchema="vocabulary")
}
}
