% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{createConceptHierarchy}
\alias{createConceptHierarchy}
\title{Create the concept hierarchy}
\usage{
createConceptHierarchy(connectionDetails, resultsDatabaseSchema,
  vocabDatabaseSchema, sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}
}
\description{
Create the concept hierarchy
}
\details{
Post-processing, create the concept hierarchy
}
