% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{createIndices}
\alias{createIndices}
\title{Create indicies}
\usage{
createIndices(connectionDetails, resultsDatabaseSchema, sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}
}
\description{
Create indicies
}
\details{
Post-processing, create indices to help performance. Cannot be used with Redshift.
}
