% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{dropAllScratchTables}
\alias{dropAllScratchTables}
\title{Drop all the scratch tables}
\usage{
dropAllScratchTables(connectionDetails, scratchDatabaseSchema,
  tempAchillesPrefix = "tmpach", tempHeelPrefix = "tmpheel",
  numThreads = 1)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{scratchDatabaseSchema}{string name of database schema that Achilles scratch tables were written to.}

\item{tempAchillesPrefix}{The prefix to use for the "temporary" (but actually permanent) Achilles analyses tables. Default is "tmpach"}

\item{tempHeelPrefix}{The prefix to use for the "temporary" (but actually permanent) Heel tables. Default is "tmpheel"}

\item{numThreads}{The number of threads to use to run this function. Default is 1 thread.}
}
\description{
Drop all the scratch tables
}
\details{
Drop all possible Achilles and Heel scratch tables
}
