% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportAO.R, R/exportAres.R
\name{exportAO}
\alias{exportAO}
\title{exportAO}
\usage{
exportAO(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema,
  outputPath
)

exportAO(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema,
  outputPath
)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Name of the database schema that contains the OMOP CDM.}

\item{resultsDatabaseSchema}{Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

See \code{showReportTypes} for a list of all report types}

\item{outputPath}{A folder location to save the JSON files. Default is current working folder}
}
\value{
none

none
}
\description{
\code{exportAO} Exports Achilles statistics - alpha option

\code{exportAO} Exports Achilles statistics - alpha option
}
\details{
Creates export files
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="yourserver")
  exportToJson(connectionDetails, cdmDatabaseSchema="cdm4_sim", outputPath="your/output/path")
}
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="yourserver")
  exportToJson(connectionDetails, cdmDatabaseSchema="cdm4_sim", outputPath="your/output/path")
}
}
