% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportAres.R
\name{exportAres}
\alias{exportAres}
\title{exportAres}
\usage{
exportAres(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  outputPath,
  vocabDatabaseSchema = cdmDatabaseSchema
)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Name of the database schema that contains the OMOP CDM.}

\item{resultsDatabaseSchema}{Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema}

\item{outputPath}{A folder location to save the JSON files. Default is current working folder}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

See \code{showReportTypes} for a list of all report types}
}
\value{
none
}
\description{
\code{exportAres} Exports Achilles statistics for use with the Ares interface
}
\details{
Creates export files for the Ares interface
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="yourserver")
  exportToJson(connectionDetails, cdmDatabaseSchema="cdm4_sim", outputPath="your/output/path")
}
}
