% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToAres.R
\name{exportToAres}
\alias{exportToAres}
\title{exportToAres}
\usage{
exportToAres(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema,
  outputPath,
  outputFormat = "default",
  reports = c()
)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Name of the database schema that contains the OMOP CDM.}

\item{resultsDatabaseSchema}{Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{outputPath}{A folder location to save the JSON files. Default is current working folder}

\item{outputFormat}{default or alternatively "duckdb" to use parquet and duckdb formats.}

\item{reports}{vector of reports to run, c() defaults to all reports

See \code{showReportTypes} for a list of all report types}
}
\value{
none
}
\description{
\code{exportToAres} Exports Achilles statistics for ARES
}
\details{
Creates export files
}
