% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{exportVocabStats}
\alias{exportVocabStats}
\title{Full export of all vocabulary stats}
\usage{
exportVocabStats(connectionDetails, resultsDatabaseSchema = "webapi",
  outputPath = "./out", topX = 20)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is 'webapi'}

\item{outputPath}{string in this folder the mapping stats are written.}

\item{topX}{}

\item{mappingName}{Name of the mapping, e.g. 'condition', 'observation', 'observation_value', ... For a full list see \code{mappingFields.csv}.}
}
\description{
\code{exportVocabStats} is a convenience function that creates a csv file for every mapping stat.
}
\details{
\code{exportVocabStats} is a convenience function that creates a csv file for every mapping stat.
The result is one mapping overview csv and a folder for each CDM source to concept mapping type.
Each folder contains a detailed mapping statistic and the \code{topX} mapped and not mapped concepts.
}
