% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDbSummary.R
\name{generateDbSummary}
\alias{generateDbSummary}
\title{generateDbSummary}
\usage{
generateDbSummary(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  country,
  provenance
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM
schema. On SQL Server, this should specifiy both the database and the
schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final
results to. Default is cdmDatabaseSchema. On SQL Server, this should
specifiy both the database and the schema, so for example, on SQL
Server, 'cdm_results.dbo'.}

\item{country}{The country of origin of the database}

\item{provenance}{The provenance of the data (EHR, claims, registry, etc)}
}
\value{
none
}
\description{
\code{generateDbSummary} can be run after the Achilles analyses are complete
to create a high-level database summary.
}
\details{
Used to generate a high-level database summary consisting of earliest date available, latest
date available, median age at first observation, total persons, etc. This function
creates a summary table meant for a manuscript detailing the network of databases
used in an analysis
}
\examples{
\dontrun{
connectionDetails <- DatabaseConnector::createConnectionDetails(
  dbms = "sql server",
  server = "yourserver"
)
dbSummary <- generateDbSummary(connectionDetails,
  cdmDatabaseSchema = "cdm_schema",
  resultsDatabaseSchema = "results_schema",
  country = "Country of Origin",
  provenance = "Provenance of data"
)
}
}
