% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{getAnalysisDetails}
\alias{getAnalysisDetails}
\title{Get all analysis details}
\usage{
getAnalysisDetails()
}
\value{
A data.frame with the analysis details.
}
\description{
Get all analysis details
}
\details{
Get a list of all analyses with their analysis IDs and strata.
}
