% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AchillesViewResults.R
\name{launchHeelResultsViewer}
\alias{launchHeelResultsViewer}
\title{Launch the Achilles Heel Shiny app}
\usage{
launchHeelResultsViewer(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema, outputFolder)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can fetch final results from.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{outputFolder}{Path to store logs and SQL files}
}
\description{
Launch the Achilles Heel Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the Achilles Heel results
}
