% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{listAllMappings}
\alias{listAllMappings}
\title{Create table of all source to target concept mappings}
\usage{
listAllMappings(connectionDetails, resultsDatabaseSchema = "webapi",
  condition_source_vocabularies = "", drug_source_vocabularies = "",
  device_source_vocabularies = drug_source_vocabularies,
  procedure_source_vocabularies = "",
  specimen_source_vocabularies = "",
  measurement_source_vocabularies = "",
  measurement_unit_source_vocabularies = "",
  measurement_value_source_vocabularies = "",
  observation_source_vocabularies = "",
  observation_unit_source_vocabularies = measurement_unit_source_vocabularies,
  observation_qualifier_source_vocabularies = "",
  death_source_vocabularies = "", specialty_source_vocabularies = "")
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is 'webapi'}

\item{condition_source_vocabularies}{string or vector ids of the source vocabularies used to map condition_source_values to condition_concept_ids via the source_to_concept_map table.}
}
\description{
\code{createMappingOverview} creates and fills a table with concept mappings for all source/target pairs. These pairs can be found in \code{mappingFields.csv}.
The \code{_source_vocabularies} are used to retrieve the \code{source_code_description} of \code{_source_value}'s. This gives more informative overviews of the top (not) mapped concepts.
}
\details{
\code{createMappingOverview} creates and fills a table with concept mappings for all source/target pairs. These pairs can be found in \code{mappingFields.csv}.
}
