% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMissingAnalyses.r
\name{listMissingAnalyses}
\alias{listMissingAnalyses}
\title{listMissingAnalyses}
\usage{
listMissingAnalyses(connectionDetails, resultsDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that contains
achilles_results and achilles_results_dist tables.}
}
\value{
A dataframe which is a subset of \code{getAnalysisDetails}
}
\description{
\code{listMissingAnalyses} Find and return analyses that exist in \code{getAnalysisDetails}, but
not in achilles_results or achilles_results_dist
}
\examples{
\dontrun{
Achilles::listMissingAnalyses(
  connectionDetails = connectionDetails,
  resultsDatabaseSchema = "results"
)
}

}
