% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{mappingStats}
\alias{mappingStats}
\title{mappingStats}
\usage{
mappingStats(connectionDetails, resultsDatabaseSchema = "webapi",
  mappingName = "")
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is 'webapi'}

\item{mappingName}{Name of the mapping, e.g. 'condition', 'observation', 'observation_value', ... For a full list see \code{mappingFields.csv}. Default to all mappings (NULL)}
}
\value{
A dataframe with results
}
\description{
\code{mappingStats} returns mapping statistics for all or for a specific mapping
}
\details{
\code{mappingStats} returns mapping statistics for all or for a specific mapping
}
