% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{mappingStatsOverview}
\alias{mappingStatsOverview}
\title{mappingStatsOverview}
\usage{
mappingStatsOverview(connectionDetails, resultsDatabaseSchema = "webapi",
  withTotalRow = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is 'webapi'}

\item{withTotalRow}{boolean}
}
\value{
A dataframe with overview of mapping statistics
}
\description{
\code{mappingStats} returns an overview of the mappings per mapping type
}
\details{
\code{mappingStats} returns an overview of the mappings per mapping type
}
