% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumAcrossYears.r
\name{sumAcrossYears}
\alias{sumAcrossYears}
\title{For a monhtly time series, compute sum and proportion by month across all years}
\usage{
tsSum <- sumAcrossYears(tsData)
}
\arguments{
\item{tsData}{A time series object}
}
\value{
A data frame reporting the monthly sum across all years and the proportion this sum contributes to the total.
}
\description{
For a monhtly time series, compute sum and proportion by month across all years
}
