% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{topMapped_}
\alias{topMapped_}
\title{Convenience function used by both topMapped and topNotMapped}
\usage{
topMapped_(sqlFilename, connectionDetails, resultsDatabaseSchema,
  mappingName, topX)
}
\description{
Convenience function used by both topMapped and topNotMapped
}
