% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyMapping.R
\name{topNotMapped}
\alias{topNotMapped}
\title{Top Concepts Not Mapped}
\usage{
topNotMapped(connectionDetails, resultsDatabaseSchema = "webapi",
  mappingName, topX = 20)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is 'webapi'}

\item{mappingName}{Name of the mapping, e.g. 'condition', 'observation', 'observation_value', ... For a full list see \code{mappingFields.csv}.}

\item{topX}{}
}
\value{
A dataframe with the results
}
\description{
\code{topNotMapped} returns a dataframe of the \code{topX} unmapped concepts, based on occurrence
}
\details{
\code{topNotMapped} returns a dataframe of the \code{topX} unmapped concepts, based on occurrence
}
