% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsCompleteYears.r
\name{tsCompleteYears}
\alias{tsCompleteYears}
\title{Trim a monthly time series object to so that partial years are removed}
\usage{
tsData <- tsCompleteYears(tsData)
}
\arguments{
\item{tsData}{A time series object}
}
\value{
A time series with partial years removed.
}
\description{
Trim a monthly time series object to so that partial years are removed
}
\details{
This function is only supported for monthly time series
}
