% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{validateCdmSchema}
\alias{validateCdmSchema}
\title{Validate the CDM schema}
\usage{
validateCdmSchema(connectionDetails, cdmDatabaseSchema, runCostAnalysis,
  cdmVersion, sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{string name of database schema that contains OMOP CDM. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.}

\item{runCostAnalysis}{Boolean to determine if cost analysis should be run. Note: only works on CDM v5 and v5.0.1 style cost tables.}

\item{cdmVersion}{Define the OMOP CDM version used:  currently supports "5", "5.0.1".  Default = "5". v4 support is in \code{\link{achilles_v4}}.}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}
}
\description{
Validate the CDM schema
}
\details{
Runs a validation script to ensure the CDM is valid based on v5.x
}
