% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{validateSchema}
\alias{validateSchema}
\title{Validate the CDM schema}
\usage{
validateSchema(connectionDetails, cdmDatabaseSchema, runCostAnalysis,
  sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{string name of database schema that contains OMOP CDM. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.}

\item{runCostAnalysis}{Boolean to determine if cost analysis should be run. Note: only works on CDM v5 and v5.1.0+ style cost tables.}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}
}
\description{
Validate the CDM schema
}
\details{
Runs a validation script to ensure the CDM is valid based on v5.x
}
