% Generated by roxygen2 (4.0.1): do not edit by hand
\name{achilles}
\alias{achilles}
\title{achilles}
\usage{
achilles(connectionDetails, cdmSchema, resultsSchema, sourceName = "",
  analysisIds, createTable = TRUE, smallcellcount = 5)
}
\arguments{
  \item{connectionDetails}{An R object of type
  ConnectionDetail (details for the function that contains
  server info, database type, optionally username/password,
  port)}

  \item{cdmSchema}{string name of database schema that
  contains OMOP CDM and vocabulary}

  \item{resultsSchema}{string name of database schema that
  we can write results to. Default is cdmSchema}

  \item{sourceName}{string name of the database, as
  recorded in results}

  \item{analysisIds}{(optional) a vector containing the set
  of Achilles analysisIds for which results will be
  generated. If not specified, all analyses will be
  executed. See \code{data(analysesDetails)} for a list of
  all Achilles analyses and their Ids.}

  \item{createTable}{If true, new results tables will be
  created in the results schema. If not, the tables are
  assumed to already exists, and analysis results will be
  added}

  \item{smallcellcount}{To avoid patient identifiability,
  cells with small counts (<= smallcellcount) are deleted.}
}
\value{
An object of type \code{achillesResults} containing details
for connecting to the database containing the results
}
\description{
\code{achilles} creates descriptive statistics summary for
an entire OMOP CDM instance.
}
\details{
PATRICK HOMEWORK: complete details
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="RNDUSRDHIT07.jnj.com")
  achillesResults <- achilles(connectionDetails, "cdm4_sim", "scratch", "TestDB")
  fetchAchillesAnalysisResults(connectionDetails, "scratch", 106)
}
}

