% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{achilles}
\alias{achilles}
\title{The main Achilles analysis (for v5+)}
\usage{
achilles(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  scratchDatabaseSchema = resultsDatabaseSchema, sourceName = "",
  analysisIds, smallCellCount = 5, cdmVersion = "5", runHeel = TRUE,
  validateCdmSchema = FALSE, runCostAnalysis = FALSE, sqlOnly = FALSE,
  numThreads = 1, tempAchillesPrefix = "tmpach", dropScratchTables = TRUE)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetails (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM (including Vocabulary). 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{scratchDatabaseSchema}{Fully qualified name of the database schema that will store all of the intermediate scratch tables, so for example, on SQL Server, 'cdm_scratch.dbo'. 
Must be accessible to/from the cdmDatabaseSchema and the resultsDatabaseSchema. Default is resultsDatabaseSchema. 
Making this "#" will run Achilles in single-threaded mode and use temporary tables instead of permanent tables.}

\item{sourceName}{String name of the database, as recorded in results.}

\item{analysisIds}{(OPTIONAL) A vector containing the set of Achilles analysisIds for which results will be generated. 
If not specified, all analyses will be executed. Use \code{\link{getAnalysisDetails}} to get a list of all Achilles analyses and their Ids.}

\item{smallCellCount}{To avoid patient identifiability, cells with small counts (<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.}

\item{cdmVersion}{Define the OMOP CDM version used:  currently supports v5 and above.  Default = "5". v4 support is in \code{\link{achilles_v4}}.}

\item{runHeel}{Boolean to determine if Achilles Heel data quality reporting will be produced based on the summary statistics.  Default = TRUE}

\item{validateCdmSchema}{Boolean to determine if CDM Schema Validation should be run. This could be very slow.  Default = FALSE}

\item{runCostAnalysis}{Boolean to determine if cost analysis should be run. Note: only works on CDM v5 and v5.0.1 style cost tables.}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{numThreads}{(optional) Only for v5.x: The number of threads to use to run Achilles in parallel. Default is 1 thread.}

\item{tempAchillesPrefix}{(optional) The prefix to use for the scratch Achilles analyses tables. Default is "tmpach"}

\item{dropScratchTables}{In multi-threaded mode: TRUE = drop the scratch tables (may take time depending on dbms), FALSE = leave them in place for later removal.}
}
\value{
An object of type \code{achillesResults} containing details for connecting to the database containing the results
}
\description{
\code{achilles} creates descriptive statistics summary for an entire OMOP CDM instance.
}
\details{
\code{achilles} creates descriptive statistics summary for an entire OMOP CDM instance.
}
\examples{
                              \dontrun{
                                          connectionDetails <- createConnectionDetails(dbms="sql server", server="some_server")
                                          achillesResults <- achilles(connectionDetails = connectionDetails, 
                                            cdmDatabaseSchema = "cdm", 
                                            resultsDatabaseSchema="results", 
                                            scratchDatabaseSchema="scratch",
                                            sourceName="Some Source", 
                                            cdmVersion = "5.0.1", 
                                            runCostAnalysis = TRUE, 
                                            numThreads = 10)
                                        }
}
