% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{achillesHeel}
\alias{achillesHeel}
\title{execution of data quality rules}
\usage{
achillesHeel(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema, cdmVersion = "5",
  vocabDatabaseSchema = cdmDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{string name of database schema that contains OMOP CDM. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all temporary tables to be managed. Requires create/insert permissions to this database.}

\item{resultsDatabaseSchema}{string name of database schema that we can write results to. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{cdmVersion}{Define the OMOP CDM version used:  currently support "4" and "5".  Default = "4"}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}
}
\value{
nothing is returned
}
\description{
\code{achillesHeel} executes data quality rules (or checks) on pre-computed analyses (or measures).
}
\details{
\code{achillesHeel} contains number of rules (authored in SQL) that are executed againts achilles results tables.
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="myserver")
  achillesHeel <- achilles(connectionDetails, cdmDatabaseSchema="mycdm", resultsDatabaseSchema="scratch", vocabDatabaseSchema="vocabulary")
}
}
