% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataSource.R
\name{addDataSource}
\alias{addDataSource}
\title{addDataSource}
\usage{
addDataSource(jsonFolderPath, dataSourcesFilePath, dataSourceName)
}
\arguments{
\item{jsonFolderPath}{Folder path of the Json files generated by \code{exportToJson}.}

\item{dataSourcesFilePath}{The full file path where datasource file will be saved.}

\item{dataSourceName}{The human readable name of the new data source}
}
\value{
none
}
\description{
\code{addDataSource} adds a data source to the datasource.json file used by AchillesWeb.
}
\details{
Used to update the datasources file with the reference to a specified datasource. This makes the
new datasource findable for OHDSI tools. If the datasources file exists, the data source will be
added to the file. If the datasources file does not exist, a new file wil be initialized with the
specified data source.
}
\examples{
\dontrun{
jsonFolderPath <- "your/output/path"
dataSourcesFilePath <- "/path/to/datasources.json"
dataSourceName <- "My New CDM"
connectionDetails <- DatabaseConnector::createConnectionDetails(dbms = "sql server",
                                                                server = "yourserver")
exportToJson(connectionDetails,
             cdmDatabaseSchema = "cdm5",
             resultsDatabaseSchema = "results",
             outputPath = jsonFolderPath)
addDataSource(jsonFolderPath, dataSourcesFilePath, dataSourceName)
}
}
