% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles-package.R
\docType{data}
\name{analysesDetails}
\alias{analysesDetails}
\title{Achilles analyses details}
\format{A data frame with 7 columns}
\usage{
data(analysesDetails)
}
\description{
A data frame describing all Achilles analyses and their IDs. The variables are as follows:
}
\details{
\itemize{
  \item ANALYSIS_ID 
  \item ANALYSIS_NAME
  \item STRATUM_1_NAME
  \item STRATUM_2_NAME
  \item STRATUM_3_NAME
  \item STRATUM_4_NAME
  \item STRATUM_5_NAME
}
}
\keyword{datasets}
