% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{createConceptHierarchy}
\alias{createConceptHierarchy}
\title{(DEPRECATED) Create the concept hierarchy}
\usage{
createConceptHierarchy(connectionDetails, resultsDatabaseSchema,
  scratchDatabaseSchema, vocabDatabaseSchema, outputFolder,
  numThreads = 1, tempAchillesPrefix = "tmpach", sqlOnly = FALSE,
  verboseMode = TRUE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{scratchDatabaseSchema}{Fully qualified name of the database schema that will store all of the intermediate scratch tables, so for example, on SQL Server, 'cdm_scratch.dbo'. 
Must be accessible to/from the cdmDatabaseSchema and the resultsDatabaseSchema. Default is resultsDatabaseSchema. 
Making this "#" will run Achilles in single-threaded mode and use temporary tables instead of permanent tables.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{outputFolder}{Path to store logs and SQL files}

\item{numThreads}{(OPTIONAL, multi-threaded mode) The number of threads to use to run Achilles in parallel. Default is 1 thread.}

\item{tempAchillesPrefix}{(OPTIONAL, multi-threaded mode) The prefix to use for the scratch Achilles analyses tables. Default is "tmpach"}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{verboseMode}{Boolean to determine if the console will show all execution steps. Default = TRUE}
}
\description{
(DEPRECATED) Create the concept hierarchy
}
\details{
This function is no longer necessary for Atlas Data Sources, as its functionality is handled by WebAPI.
}
