% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Achilles.R
\name{createConceptHierarchy}
\alias{createConceptHierarchy}
\title{Create the concept hierarchy}
\usage{
createConceptHierarchy(connectionDetails, resultsDatabaseSchema,
  scratchDatabaseSchema, vocabDatabaseSchema, numThreads = 1,
  tempAchillesPrefix = "tmpach", sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{scratchDatabaseSchema}{Fully qualified name of the database schema that will store all of the intermediate scratch tables, so for example, on SQL Server, 'cdm_scratch.dbo'. 
Must be accessible to/from the cdmDatabaseSchema and the resultsDatabaseSchema. Default is resultsDatabaseSchema. 
Making this "#" will run Achilles in single-threaded mode and use temporary tables instead of permanent tables.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{numThreads}{(OPTIONAL, multi-threaded mode) The number of threads to use to run Achilles in parallel. Default is 1 thread.}

\item{tempAchillesPrefix}{(OPTIONAL, multi-threaded mode) The prefix to use for the scratch Achilles analyses tables. Default is "tmpach"}

\item{sqlOnly}{TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}
}
\description{
Create the concept hierarchy
}
\details{
Post-processing, create the concept hierarchy.
Please note: this table creation only requires the Vocabulary, not the CDM itself. 
You could run this once for 1 Vocab version, and then copy the table to all CDMs using that Vocab.
}
