% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exportToJson.R
\name{exportConditionEraToJson}
\alias{exportConditionEraToJson}
\title{exportConditionEraToJson}
\usage{
exportConditionEraToJson(connectionDetails, cdmSchema, resultsSchema,
  outputPath = getwd(), cdmVersion = "4")
}
\arguments{
  \item{connectionDetails}{An R object of type
  ConnectionDetail (details for the function that contains
  server info, database type, optionally username/password,
  port)}

  \item{cdmSchema}{Name of the database schema that
  contains the vocabulary files}

  \item{resultsSchema}{Name of the database schema that
  contains the Achilles analysis files. Default is
  cdmSchema}

  \item{outputPath}{A folder location to save the JSON
  files. Default is current working folder}
}
\value{
none
}
\description{
\code{exportConditionEraToJson} Exports Achilles Condition
Era report into a JSON form for reports.
}
\details{
Creates individual files for Condition Era report found in
Achilles.Web
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="yourserver")
  exportConditionEraToJson(connectionDetails, cdmSchema="cdm4_sim", outputPath="your/output/path")
}
}

