% Generated by roxygen2 (4.0.0): do not edit by hand
\name{exportDeathToJson}
\alias{exportDeathToJson}
\title{exportDeathToJson}
\usage{
exportDeathToJson(connectionDetails, cdmSchema, resultsSchema,
  outputPath = getwd())
}
\arguments{
  \item{connectionDetails}{An R object of type
  ConnectionDetail (details for the function that contains
  server info, database type, optionally username/password,
  port)}

  \item{cdmSchema}{Name of the database schema that
  contains the vocabulary files}

  \item{resultsSchema}{Name of the database schema that
  contains the Achilles analysis files. Default is
  cdmSchema}

  \item{outputPath}{A folder location to save the JSON
  files. Default is current working folder}
}
\value{
none
}
\description{
\code{exportDeathToJson} Exports Achilles Death report into
a JSON form for reports.
}
\details{
Creates individual files for Death report found in
Achilles.Web
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="yourserver")
  exportDeathToJson(connectionDetails, cdmSchema="cdm4_sim", outputPath="your/output/path")
}
}

