% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToJson.R
\name{exportMetaToJson}
\alias{exportMetaToJson}
\title{exportMetaToJson}
\usage{
exportMetaToJson(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  outputPath = getwd(), cdmVersion = "4",
  vocabDatabaseSchema = cdmDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Name of the database schema that contains the vocabulary files}

\item{resultsDatabaseSchema}{Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema}

\item{outputPath}{A folder location to save the JSON files. Default is current working folder}

\item{cdmVersion}{Define the OMOP CDM version used:  currently support "4" and "5".  Default = "4"}

\item{vocabDatabaseSchema}{string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}
}
\value{
none
}
\description{
\code{exportMetaToJson} Exports Achilles Heel report into a JSON form for reports.
}
\details{
Creates individual files for Achilles Heel report found in Achilles.Web
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="yourserver")
  exportMetaToJson(connectionDetails, cdmDatabaseSchema="cdm4_sim", outputPath="your/output/path")
}
}
