% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AchillesViewResults.R
\name{fetchAchillesAnalysisResults}
\alias{fetchAchillesAnalysisResults}
\title{fetchAchillesAnalysisResults}
\usage{
fetchAchillesAnalysisResults(connectionDetails, resultsDatabaseSchema,
  analysisId)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can fetch final results from.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{analysisId}{A single analysisId}
}
\value{
An object of type \code{achillesAnalysisResults}
}
\description{
\code{fetchAchillesAnalysisResults} returns the results for one Achilles analysis Id.
}
\details{
See \code{data(analysesDetails)} for a list of all Achilles analyses and their Ids.
}
\examples{
                              \dontrun{
                                           connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="myserver")
                                           achillesResults <- achilles(connectionDetails, "cdm4_sim", "scratch", "TestDB")
                                           fetchAchillesAnalysisResults(connectionDetails, "scratch",106)
                                        }
}
