% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AchillesViewResults.R
\name{fetchAchillesAnalysisResults}
\alias{fetchAchillesAnalysisResults}
\title{fetchAchillesAnalysisResults}
\usage{
fetchAchillesAnalysisResults(connectionDetails, resultsDatabase, analysisId)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabase}{Name of database containing the Achilles descriptive statistics.}

\item{analysisId}{A single analysisId}
}
\value{
An object of type \code{achillesAnalysisResults}
}
\description{
\code{fetchAchillesAnalysisResults} returns the results for one Achilles analysis Id.
}
\details{
See \code{data(analysesDetails)} for a list of all Achilles analyses and their Ids.
}
\examples{
\dontrun{
  connectionDetails <- DatabaseConnector::createConnectionDetails(dbms="sql server", server="myserver")
  achillesResults <- achilles(connectionDetails, "cdm4_sim", "scratch", "TestDB")
  fetchAchillesAnalysisResults(connectionDetails, "scratch",106)
}
}
