% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStationary.r
\name{isStationary}
\alias{isStationary}
\title{Determine whether or not a time series is stationary in the mean}
\usage{
isStationary(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A boolean indicating whether or not the given time series is stationary.
}
\description{
Uses the Augmented Dickey-Fuller test to determine when the time series has a unit root.
}
\details{
A time series must have a minimum of three complete years of data.
For details on the implementation of the Augmented Dickey-Fuller test, 
see the tseries package on cran.
}
